Move viking states into their own folder
