+++ Assets/00_Content/Scripts/Vikings/States/DesiringVikingState.cs
@@ -1,7 +1,7 @@
 ﻿using Interactables;
 using UnityEngine;
 
-namespace Vikings {
+namespace Vikings.States {
 	public class DesiringVikingState : VikingState {
 		public DesiringVikingState(Viking viking) : base(viking) {
 			viking.bodyMeshRenderer.material = viking.desiringMaterial;
